add skills effect factory

+++ Assets/Code/Skills/CharacterSkill/Core/EffectsCore/IBuffSkillEffect.cs
@@ -4,7 +4,7 @@ namespace Code.Skills.CharacterSkill.Core.EffectsCore
 {
 public interface IBuffSkillEffect : ISkillEffect
 {
-	public float BuffDuration { get; }
+	public int BuffDuration { get; }
 	public int BuffAmount { get; }
 }
 }
